requirements2.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.7
fdb==2.0.1
future==0.18.2
idna==2.8
mysqlclient==1.4.6
pkg-resources==0.0.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
